Redesign and refactor the map file code; begin map editor implementation.
The newly redesigned map file format should allow the map engine to be
much more powerful.  Depth=0 is no longer reserved for players and items;
instead, the maps truly have depth, complete with players and items being
able to be placed on any layers.  This will allow for a map to contain
constructs such as balconies, etc. which are independent of the surface
below them.  Very powerful.  Oh, and the code is simpler, too.  Win-win.

The existing map code has also been refactored to make it easier to work
with.