Map rendering, updated map format, etc.
The map renderer is now integrated with the map editor.  Maps are rendered
in the map editor when open (however, they cannot actually be edited yet).
The map file format has significantly changed; most significantly, maps
now have finite depth (specified in the header) and unique tiles for each
layer of depth.  In addition, the Layer class no longer exists; everything
is dealt with as tiles having a z-coordinate.  This significantly simplifies
the map engine.