Fix the circular dependency problem.
Yes, I had to revert to that backup commit I made. We're going to
run into more of these circular dependencies in the future - the
easiest fix is to just pass down references to the parent main
window throughout the class chain.  That's what I did for the
EditTools.Toolbox class just now.