Pen tool is starting to work. A few bugs still.
Most notably the undo/redo features for the pen still aren't
working.  I would fix them now, but again, I'm tired and want to
go to sleep.  I'll take care of them in the morning... maybe.

Note to self: Add keyboard shortcuts for moving up and down
through the layers of a map.