Map editor save/load works.  Updated icons, fixed credits.
We've got some new icons for the map editor now.  They're from
the GNOME icon theme; credits are in the CREDITS file (used to
be the AUTHORS file).

Save/load of maps is now working.  I still have to implement
the rectangle draw and flood bucket tools, and get the subtle
parts of the main interface working properly.  With that said,
we now have a basic working map editor, and work on the client
and server can proceed.  We'll be using Python's asyncore module
for networking.  First draft of the protocol coming soon.