Start work on the network protocol.  And other stuff.
Big change in the client this time: no more standard tileset. The
one we were using before was never particularly clear about its
license, and even though it was made freely available online for
use in games, I removed it from the repository to stay on the
safe side.  So you'll have to supply your own tileset now if you
want to test anything graphical.

We'll add in a standard tileset later on, probably under a
CC-BY-SA license (but who knows what we'll end up going with).
But hey, hopefully we can recruit a pixel artist at some point
and get some custom artwork for the engine.  That's for later.