Start implementing server; going to use eventlet instead of gevent.
Started implementing the server using gevent for asynchronous networking, but
gevent is ridiculously unstable (I get *segfaults* when I run the *example
code*), so I'm going to switch to eventlet.  This is really just a backup
commit in case I (for some unknown reason) decide to go back to gevent.

Also, gevent's documentation is kinda crappy.  Between the bad docs and the
segfaults all over the place, I don't want to deal with it anymore.