Start actual work on the client; major directory restructure.
The new directory structure clearly separates each component into
a separate Python package with its own setup.py file.  The PyDev
project has also been updated to reflect the new PYTHONPATH needed
for this structure.

Work on the client is coming along nicely.  We have an "operational"
server (the quotes because it doesn't do anything once a connection
is made yet), but as soon as the client's networking is up and
running, we'll set up and test the connection negotiation code. From
there, it's login/registration code, and then... maps and character
movement!  And then development will really start to pick up.

I haven't had a chance to give the new directory structure a good
test, so there's probably some broken stuff.  I'll fix that soon.