Refactor the sprite code to handle multiple server resource folders.
At some point I want to overhaul the sprite system and get rid of
those ".meta" files.  I'd like to move to a more flexible XML
format for describing the contents of the spriteset.  Oh, and I
still need to actually test the animation code, even though that
was written months ago.  Maybe I'll just overhaul it all along
with the .meta files.