More bugfixes and basic features in the map editor.
This update comes with a backwards-incompatible change to the
map file format.  These updates will stop becoming backwards
incompatible soon; however, we're so early in development that
it would be pointless to waste the effort of keeping old maps
loadable (since the client doesn't even run yet - but we're
not far away from that).  Basically, the ability to set individual
tiles as "blocked" has been removed; this no longer makes sense
since the rendering (and movement) of players is restricted to
the layer specified by the "player depth" field in the map header.
So, in a future update, there will be a new data section of the
map files after the existing Tiles section which keeps track of
"positional flags."  These will define map behavior at a given
(x,y) coordinate for all depths and will include properties such
as whether a tile is impassable.

One other update to the map format this time around: It's now
possible to set a string in the map header telling the client
what background image to display behind the map.  Useful for...
I don't know, mountain summits overlooking the night sky or some
other "scenic" map for your game.  I had meant to include this
from the beginning, and I guess it just slipped my mind when I
wrote the original map format specification.  Anyway, it's added
now.