Client startup interface added; network integration started.
We have a new temporary background image for the client.  It's a
fractal image that I generated in GIMP.  I kinda like it.  We'll
replace it eventually, but for prototyping the engine it looks nice.

More of the default style has been defined.  There's a lot of work
left to do on it in order to keep a unified appearance across all
platforms, but we're off to a good start.  I'm excited about the use
of Qt stylesheets to enable server developers to customize the look
and feel of clients connected to their server; I think this is a
cool feature unique to the xVector engine, and it's pretty awesome.

Speaking of UI stuff, the client has a UI now.  Well, a startup UI,
anyway.  The settings screen is still missing, mostly because I
haven't decided what kinds of settings I want to expose to the user
yet.  There's probably going to be something about framerate
limiting (early tests of unlimited framerates have... interesting
effects; it might be best to have an absolute cap around 60fps
or so).  But that will come later.

Oh, and the client and server actually talk to each other now.
Sort of.  The client can connect and begin a connection negotiation,
but the server doesn't actually do anything with that yet.  That
comes next!

Oh, right - the metaserver is working now.  I'm going to host a
centralized metaserver for the engine to make it easy for developers
to share their games with others; you'll be able to boot up the
client and log into any of a whole list of public servers, each with
their own custom content.  And, of course, there's still a "Private
Server" option for those who would rather connect by IP and port.
They're both already working, including the server-side
implementation of the metaserver which detects when servers go
offline and hides them from the public list.  Awesomeness.