Server IPv6 support, DATABASES!, rewritten configuration file code.
Lots of changes this time around.  The server can run on both IPv4 and IPv6
now (either one or both); IPv6 support has yet to be added to the client.  The
server's configuration file parser has been rewritten (and moved to xVLib; it
will be reused in the client as well) to use a "nicer" XML format now (and a
faster parser, not that it really matters).  Database support added to the
server using SQLAlchemy, along with our first table: the IP ban records.  The
bans have yet to actually be implemented, but the database table for them is
there.