Start adding the login code, etc.
The login protocol was redesigned to address a few security issues.
The login process now runs on a separate SSL-wrapped socket which
hands over a single-use authentication token for identifying the client
to the server.  It prevents the user's password from being sent in
plaintext over the wire while at the same time keeping the database record
of password hashes relatively useless to an attacker.