Add cx_Freeze support to the map editor so it can build as EXE/ELF/whatever.
An important note about client and map editor behavior: If you're using either
as the native Python scripts (rather than a "frozen" executable), you need to
make sure that they aren't located in a subdirectory of a folder called "library.zip".
That really shouldn't ever be a problem.  The reason for this is that ClientPaths.py
checks for it as a higher-level directory in the currently executing script to
determine whether the program is being run as a native Python script or as a "frozen"
executable.  So, don't name your directories "library.zip".  (As for why you ever
would, I have no idea.)