Slight cleanup/restructuring, and migrate the docs to Sphinx.
The documentation (except for the API - that's still epydoc-generated) is now
in rst format, which means that Sphinx can be used to build pretty HTML and
PDF manuals for the software.  It's good to start early with this stuff so
that documentation isn't such a huge job later, and it's more complete.  Also
added some documentation on all of the server configuration options that have
been implemented so far.